Add additional features enabling to select the number of options and select destination having positive number of terminals
