+++ Procfile
@@ -1,1 +1,1 @@
- web: sh setup.sh && streamlit run tests/front.py
+ web: sh setup.sh && streamlit run ev_Stations_Website/front.py
