--- tests/info_stations.csv
@@ -1,92 +1,0 @@
-s_id,longitude,latitude
-S1,2.354089,48.855667
-S10,2.397724,48.86424
-S11,2.28308,48.861016
-S12,2.336857,48.824388
-S13,2.335728,48.88264
-S14,2.317465,48.840977
-S15,2.3748774000000594,48.846295
-S16,2.33891,48.889731
-S17,2.4112383000000364,48.8452653
-S18,2.268339,48.846761
-S19,2.392178,48.877409
-S2,2.355064,48.878897
-S20,2.319768,48.882718
-S21,2.3581455999999434,48.8260387
-S22,2.298576,48.841529
-S23,2.377938800000038,48.86535800000001
-S24,2.375639,48.841742
-S25,2.34066,48.89311
-S26,2.343436,48.865311
-S27,2.34046,48.86003
-S28,2.359043,48.864719
-S29,2.34556,48.854057
-S3,2.286403,48.876162
-S30,2.345032,48.845715
-S31,2.345493,48.85104
-S32,2.35242,48.84477
-S33,2.3414834000000155,48.8524588
-S34,2.331345899999974,48.85212550000001
-S35,2.307126,48.849911
-S36,2.30437,48.857552
-S37,2.327753,48.857124
-S38,2.309952,48.847931
-S39,2.31368,48.852665
-S4,2.289894,48.871465
-S41,2.303269,48.86937
-S42,2.309542,48.875854
-S45,2.379874,48.858952
-S46,2.401736,48.847485
-S47,2.365661,48.820007
-S48,2.363025,48.832893
-S49,2.320052,48.831448
-S5,2.315463,48.865501
-S50,2.325895,48.83086
-S51,2.323891,48.841198
-S52,2.31159,48.86154
-S53,2.286912,48.835709
-S56,2.3006427999999914,48.828454
-S57,2.297463,48.885689
-S58,2.302442,48.879654
-S59,2.286938,48.883282
-S6,2.382705,48.882794
-S60,2.405196,48.874134
-S62,2.321369,48.880698
-S63,2.33321,48.870628
-S64,2.342244,48.865788
-S65,2.339865,48.84824
-S66,2.326367,48.851025
-S67,2.368851,48.878799
-S68,2.360103,48.875465
-S69,2.38804,48.847446
-S7,2.3868801999999505,48.8534452
-S70,2.405248,48.834084
-S71,2.343366,48.826504
-S72,2.343626,48.82085
-S74,2.305529,48.834705
-S75,2.279884,48.846214
-S76,2.27153,48.863602
-S77,2.259932,48.837055
-S78,2.258213,48.846653
-S79,2.313751,48.881168
-S8,2.3699612000000343,48.8847205
-S80,2.352758,48.897671
-S81,2.327102,48.891102
-S82,2.371768,48.889011
-S83,2.398476,48.880421
-S84,2.41177,48.863766
-S85,2.377671,48.872192
-S86,2.35423,48.89061
-S87,2.36554,48.865097
-S88,2.367798,48.82312
-S89,2.388149,48.88842
-S9,2.293606,48.878965
-S91,2.403802,48.857094
-S92,2.286931,48.841663
-S93,2.273944,48.859184
-S94,2.389173,48.833603
-S95,2.295469,48.868511
-S96,2.34965,48.835499
-S97,2.397822,48.839947
-S98,2.36192,48.84998
-S99,2.39182,48.8778
